fix [VIDL-259]: Create alias path for components, change the logic PromoStore, create MainLayout, format code
