+++ src/layouts/MainLayout/Layout.tsx
@@ -1,8 +1,8 @@
 import { Outlet } from 'react-router';
 import { Box } from '@mui/material';
 
-import PageContent from './PageContent/PageContent';
-import Sidebar from './SidebarComponents/Sidebar';
+import PageContent from './PageContent';
+import Sidebar from './Sidebar';
 
 const Layout = () => {
   return (
