+++ vite.config.ts
@@ -1,7 +1,6 @@
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
 
-// https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
   resolve: {
@@ -12,6 +11,7 @@ export default defineConfig({
       forms: '/src/forms',
       types: '/src/types',
       layouts: '/src/layouts',
+      components: '/src/components',
     },
   },
 });
